projects/fedlab_parser/requirements.txt
# Core
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
keyring>=24.0.0

# Async
aiohttp>=3.9.0
aiogram>=3.0.0

# Data
pydantic>=2.0.0
psycopg2-binary>=2.9.0

# Parsing
playwright>=1.40.0
lxml>=5.0.0

# Embeddings & Vector DB
sentence-transformers>=2.2.0
qdrant-client>=1.7.0
chromadb>=0.5.0
numpy>=1.24.0
torch>=2.0.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Utils
tiktoken>=0.5.0
unstructured>=0.10.0
pdfplumber>=0.10.0